allows group masks
When setting group masks via arrays, we fail because attributes that are
not defaulted to true, are defaulted to false, wiping out each other.

Instead, don't set a default of false when the default it not set

e.g.

  has_bitmask_attributes :roles, 'with_role_' do |config|
    config.attribute 'manager',  0b00001
    config.attribute 'operator', 0b00010
    config.attribute 'customer', 0b00100

    config.attribute 'staff_group', 0b00011
  end

MyModel.new(roles: [:staff_group]) # the values are clobbered without the fix (maybe... hash order)